Service 계층 수정

+++ src/main/java/com/codeit/discodeit/dto/binary_contents_dto/BinaryContentDto.java
@@ -5,6 +5,6 @@ import java.util.UUID;
 public record BinaryContentDto(
     UUID id,
     String fileName,
-    Integer size,
+    long size,
     String contentType
 ) {}
